refactor: move parseTransaction to two-chain-setup
